Test: Ensure trivias are properly ignored for defining names
Gnattest uses the text of some nodes both for generating stubs and
test procedure names, but also when generating marker identifiers and
hashes. In particular, if a defining name contains some trivia tokens,
this could lead to invalid sources being generated.

This is fixed by explicitely ignoring trivia tokens when getting the
text associated with a defining name node.