test-actions: Fix processing of --test-root
The --tests-root switch is supposed to replicate the source directory
structure to re-create tests, starting from the common directory for all
sources. The subtlety is that if the path passed to --test-root is relative,
it is interpreted relative to each project's object directory, and processing
should remain local to a project (i.e. the root path component for a given
project should only be computed using the sources and source dirs of the project.

The processing of that option did not systematically reset the computed root
when changing the project being processed, resulting in spurious paths components
being included in the generated test directories.

This fixes it, and adds some context as to what the code is meant to do.